audio_utils: reduce nonlinearities in echo cancellation reference
The short tolerance of 0.0625ms used to keep the reference signal
aligned with the mic signal is too sensitive and introduces a
number of discontinuities into the reference signal for echo
cancellation.
This change opens the tolerance to 10ms, changes the location
of the inserted adjustment samples to be prepended to the signal
instead of inserted in the signal, and adds some robustness
to check that the delay estimate is accurate.

Change-Id: I923e555f905d31060b5f7c16a9cd64bb198ce2fd
Signed-off-by: Bryan Honza <honza@ti.com>
Signed-off-by: Barry Woodward <a0270655@ti.com>